refactor: move virtualization version validation to cluster
