+++ internal/cluster/validation/virtualization_test.go
@@ -1,4 +1,4 @@
-package accelerator
+package validation
 
 import (
 	"testing"
