+++ tests/e2e/accelerator_virtualization_test.go
@@ -9,7 +9,7 @@ import (
 	. "github.com/onsi/gomega"
 
 	v1 "github.com/neutree-ai/neutree/api/v1"
-	"github.com/neutree-ai/neutree/internal/accelerator"
+	clustervalidation "github.com/neutree-ai/neutree/internal/cluster/validation"
 )
 
 const (
@@ -121,7 +121,7 @@ var _ = Describe("K8s Accelerator Virtualization", Ordered,
 func requireAcceleratorVirtualizationProfile() {
 	requireImageRegistryProfile()
 
-	supported, err := accelerator.SupportsVirtualizationClusterVersion(profileClusterVersion())
+	supported, err := clustervalidation.SupportsVirtualizationClusterVersion(profileClusterVersion())
 	if err != nil {
 		Skip(fmt.Sprintf("Cluster version %q is invalid for accelerator virtualization: %v",
 			profileClusterVersion(), err))
